fix(dashboard comment system)

+++ client/src/pages/Dashboard.js
@@ -12,6 +12,7 @@ import {
   Heading,
   Button,
   Avatar,
+  FormControl,
   useToast,
   Image,
   Center,
@@ -32,7 +33,7 @@ import {
   LineElement,
 } from "chart.js";
 import { Line } from "react-chartjs-2";
-import { deleteComment } from "../utils/data/CommentData";
+import { deleteComment, updateComment } from "../utils/data/CommentData";
 
 const Dashboard = () => {
   ChartJS.register(
@@ -130,12 +131,16 @@ const Dashboard = () => {
   };
 
   const commentsDeleteFunc = () => {
-    commentDeleteList.map(async (comment) => {
-      await deleteComment(comment.course, comment.lesson.slug, comment._id)
+    if (commentDeleteList.length > 0) {
+      const promise = commentDeleteList.map(async (comment) => {
+        await deleteComment(comment.course, comment.lesson.slug, comment._id);
+      });
+
+      Promise.all(promise)
         .then(() =>
           toast({
             title: "Success",
-            description: `Comment Deleted Successfully`,
+            description: `Comments Deleted Successfully`,
             status: "success",
             duration: 5000,
             isClosable: true,
@@ -144,10 +149,54 @@ const Dashboard = () => {
         .catch((error) => {
           setErrors([...errors, error]);
         });
-    });
+    }
   };
 
-  const handleCommentsSubmit = () => {};
+  const handleCommentsSubmit = () => {
+    const promises = commentTextList.map(async (comment) => {
+      const formDataText = new FormData();
+      formDataText.append("text", comment.text);
+
+      // Güncellenmiş yorumu almak için updateComment fonksiyonunu bekleyin
+      const updatedComment = await updateComment(
+        comment.course,
+        comment.lesson.slug,
+        comment.commentId,
+        formDataText
+      );
+
+      return updatedComment; // Güncellenmiş yorumu promise zinciri boyunca geçirin
+    });
+
+    Promise.all(promises)
+      .then((updatedComments) => {
+        // Tüm güncellenmiş yorumları aldıktan sonra commentTextList'i güncelleyin
+        const updatedCommentTextList = commentTextList.map((comment) => {
+          const updatedComment = updatedComments.find(
+            (updatedComment) => updatedComment._id === comment.commentId
+          );
+          if (updatedComment) {
+            return { ...comment, text: updatedComment.text };
+          }
+          return comment;
+        });
+
+        setCommentTextList(updatedCommentTextList);
+
+        // Başarılı tost mesajını gösterin
+        toast({
+          title: "Success",
+          description: `Comments Update Successfully`,
+          status: "success",
+          duration: 5000,
+          isClosable: true,
+        });
+      })
+      .catch((error) => {
+        // Hata durumunda hatayı kaydedin
+        setErrors([...errors, error]);
+      });
+  };
 
   useEffect(() => {
     setOwnedCourses(
@@ -176,14 +225,20 @@ const Dashboard = () => {
     );
   }, [ownedCourses]);
 
-  useEffect(() => {
-    setActiveButtonIndices(Array(comments.length).fill(false));
-
+  const updatedCommentTextListFunc = () => {
     const updatedCommentTextList = comments.map((comment) => ({
       commentId: comment?._id,
       text: comment.text,
+      lesson: comment.lesson,
+      course: comment.course,
     }));
     setCommentTextList(updatedCommentTextList);
+  };
+
+  useEffect(() => {
+    setActiveButtonIndices(Array(comments.length).fill(false));
+
+    updatedCommentTextListFunc();
   }, [comments]);
 
   useEffect(() => {
@@ -513,6 +568,7 @@ const Dashboard = () => {
                 : ""}
             </Flex>
           </GridItem>
+
           <GridItem
             rowSpan={20}
             colSpan={1}
@@ -527,147 +583,164 @@ const Dashboard = () => {
             bgColor={"var(--bg-color)"}
             border={"2px dashed var(--secondary-color)"}
           >
-            <form onSubmit={handleCommentsSubmit()}>
-              <Flex align={"center"} justify={"space-between"} mr={"1em"}>
-                <Text fontWeight={"500"} fontSize={responsive("", "md", "lg")}>
-                  Your Comments
-                </Text>
-                <ButtonGroup>
-                  {commentsEdit && (
-                    <Button
-                      variant={"outline"}
-                      onClick={() => {
-                        setCommentsEdit(!commentsEdit);
-                        commentsDeleteFunc();
-                      }}
-                      bgColor={"var(--secondary-color)"}
-                      type="submit"
-                      color={"white"}
-                      fontSize={responsive("", "sm", "md")}
-                      border={"1px solid var(--secondary-color)"}
-                      _hover={{
-                        bgColor: "white",
-                        color: "var(--secondary-color)",
-                      }}
-                    >
-                      Save
-                    </Button>
-                  )}
+            <Flex align={"center"} justify={"space-between"} mr={"1em"}>
+              <Text fontWeight={"500"} fontSize={responsive("", "md", "lg")}>
+                Your Comments
+              </Text>
+              <ButtonGroup>
+                {commentsEdit && (
                   <Button
                     variant={"outline"}
                     onClick={() => {
                       setCommentsEdit(!commentsEdit);
-                      setActiveButtonIndices(
-                        Array(comments.length).fill(false)
-                      );
-                      setCommentDeleteList([]);
+                      commentsDeleteFunc();
+                      handleCommentsSubmit();
                     }}
-                    bgColor={"var(--accent-color)"}
+                    bgColor={"var(--secondary-color)"}
+                    type={"submit"}
                     color={"white"}
                     fontSize={responsive("", "sm", "md")}
-                    border={"1px solid var(--accent-color)"}
+                    border={"1px solid var(--secondary-color)"}
                     _hover={{
                       bgColor: "white",
-                      color: "var(--accent-color)",
+                      color: "var(--secondary-color)",
                     }}
                   >
-                    {commentsEdit ? "Reset" : "Edit"}
+                    Save
                   </Button>
-                </ButtonGroup>
-              </Flex>
-              <Flex
-                pr={"1em"}
-                overflow={"auto"}
-                mt={responsive("", ".5em", ".5em")}
-                flexDir={"column"}
-                gap={"1em"}
-              >
-                {comments &&
-                  comments.length > 0 &&
-                  comments.map(
-                    (comment, index) =>
-                      !commentDeleteList.includes(comment) && (
-                        <Flex
-                          bgColor="white"
-                          p={responsive("", ".5em", "1em")}
-                          borderRadius="10px"
-                          justify="space-between"
-                          transition="all .3s ease"
-                          key={index}
-                          flexDir="column"
-                          gap="1em"
-                        >
-                          <Flex align="center" justify="space-between">
-                            <Flex gap=".5em" align="center">
-                              <Avatar
-                                src={`http://localhost:5000/${comment.user.image}`}
-                                bgColor="var(--secondary-color)"
-                                name={comment.user?.username}
-                                size={responsive("", "sm", "sm")}
-                              />
-                              <Text
-                                fontWeight="500"
-                                fontSize={responsive("", "sm", "md")}
-                                opacity={0.9}
-                              >
-                                {comment.user?.username}
-                              </Text>
-                            </Flex>
-                            {commentsEdit && (
-                              <Button
-                                variant="outline"
-                                p=".5em"
-                                minH="max-content"
-                                minW="max-content"
-                                onClick={() =>
-                                  handleCommentButtonClick(index, comment)
-                                }
-                                fontSize={responsive("", "sm", "md")}
-                                _hover={{
-                                  bgColor: activeButtonIndices[index]
-                                    ? "var(--accent-color)"
-                                    : "unset",
-                                }}
-                                bgColor={
-                                  activeButtonIndices[index]
-                                    ? "var(--accent-color)"
-                                    : "unset"
-                                }
-                              >
-                                <i
-                                  className="fi fi-rr-trash"
-                                  style={{ position: "relative", top: "2px" }}
-                                />
-                              </Button>
-                            )}
-                          </Flex>
-
-                          {commentsEdit ? (
-                            commentTextList.map((comment) => (
-                              <Textarea
-                                key={comment._id}
-                                value={comment.text}
-                                onChange={(e) =>
-                                  setCommentTextList([
-                                    ...commentTextList,
-                                    {
-                                      commentId: comment?._id,
-                                      text: e.target.value,
-                                    },
-                                  ])
-                                }
-                              ></Textarea>
-                            ))
-                          ) : (
+                )}
+                <Button
+                  variant={"outline"}
+                  onClick={() => {
+                    setCommentsEdit(!commentsEdit);
+                    setActiveButtonIndices(Array(comments.length).fill(false));
+                    setCommentDeleteList([]);
+                    commentsEdit && updatedCommentTextListFunc();
+                  }}
+                  bgColor={"var(--accent-color)"}
+                  color={"white"}
+                  fontSize={responsive("", "sm", "md")}
+                  border={"1px solid var(--accent-color)"}
+                  _hover={{
+                    bgColor: "white",
+                    color: "var(--accent-color)",
+                  }}
+                >
+                  {commentsEdit ? "Reset" : "Edit"}
+                </Button>
+              </ButtonGroup>
+            </Flex>
+            <Flex
+              maxH={"100%"}
+              overflow={"auto"}
+              pr={"1em"}
+              h={"100%"}
+              mt={responsive("", ".5em", ".5em")}
+              flexDir={"column"}
+              gap={"1em"}
+            >
+              {comments &&
+                comments.length > 0 &&
+                comments.map((comment, index) => {
+                  const updatedComment = commentTextList.find(
+                    (c) => c.commentId === comment._id
+                  );
+                  return (
+                    !commentDeleteList.includes(comment) && (
+                      <Flex
+                        key={index}
+                        bgColor="white"
+                        p={responsive("", ".5em", "1em")}
+                        borderRadius="10px"
+                        justify="space-between"
+                        transition="all .3s ease"
+                        flexDir="column"
+                        gap="1em"
+                      >
+                        <Flex align="center" justify="space-between">
+                          <Flex gap=".5em" align="center">
+                            <Avatar
+                              src={`http://localhost:5000/${comment.user.image}`}
+                              bgColor="var(--secondary-color)"
+                              name={comment.user?.username}
+                              size={responsive("", "sm", "sm")}
+                            />
                             <Text
-                              p=".5em 1em"
-                              border="2px dashed #cfcfcf"
-                              borderRadius="10px"
+                              fontWeight="500"
+                              fontSize={responsive("", "sm", "md")}
+                              opacity={0.9}
                             >
-                              {comment.text}
+                              {comment.user?.username}
                             </Text>
+                          </Flex>
+                          {commentsEdit && (
+                            <Button
+                              variant="outline"
+                              p=".5em"
+                              minH="max-content"
+                              minW="max-content"
+                              onClick={() =>
+                                handleCommentButtonClick(index, comment)
+                              }
+                              fontSize={responsive("", "sm", "md")}
+                              _hover={{
+                                bgColor: activeButtonIndices[index]
+                                  ? "var(--accent-color)"
+                                  : "unset",
+                              }}
+                              bgColor={
+                                activeButtonIndices[index]
+                                  ? "var(--accent-color)"
+                                  : "unset"
+                              }
+                            >
+                              <i
+                                className="fi fi-rr-trash"
+                                style={{ position: "relative", top: "2px" }}
+                              />
+                            </Button>
                           )}
-                          <Flex align="center" justify="space-between">
+                        </Flex>
+
+                        {commentsEdit ? (
+                          <Textarea
+                            border={"2px dashed #cfcfcf"}
+                            value={updatedComment ? updatedComment.text : ""}
+                            _focus={{
+                              boxShadow: "none",
+                              border: "2px dashed #cfcfcf",
+                            }}
+                            onChange={(e) => {
+                              const updatedCommentTextList = [
+                                ...commentTextList,
+                              ];
+                              const commentIndex =
+                                updatedCommentTextList.findIndex(
+                                  (c) => c.commentId === comment._id
+                                );
+                              if (commentIndex !== -1) {
+                                updatedCommentTextList[commentIndex].text =
+                                  e.target.value;
+                                setCommentTextList(updatedCommentTextList);
+                              }
+                            }}
+                          ></Textarea>
+                        ) : (
+                          <Textarea
+                            readOnly={true}
+                            p=".5em 1em"
+                            border="2px dashed #cfcfcf"
+                            borderRadius="10px"
+                            _focus={{
+                              boxShadow: "none",
+                              border: "2px dashed #cfcfcf",
+                            }}
+                          >
+                            {comment.text}
+                          </Textarea>
+                        )}
+                        <Flex align="center" justify="space-between">
+                          <Flex align={"center"} gap={"1em"}>
                             <Text
                               fontWeight="500"
                               fontSize={responsive("", "sm", "md")}
@@ -675,19 +748,34 @@ const Dashboard = () => {
                             >
                               {comment.lesson?.title}
                             </Text>
-                            <Text
-                              fontWeight="500"
-                              opacity={0.9}
+                            <Flex
                               fontSize={responsive("", "sm", "md")}
+                              fontWeight={500}
+                              opacity={0.9}
+                              color={"var(--accent-color)"}
+                              gap={".3em"}
+                              align={"center"}
                             >
-                              {getDate(comment.createdAt)}
-                            </Text>
+                              <Text>{comment.point}</Text>
+                              <StarIcon
+                                pos={"relative"}
+                                bottom={"2px"}
+                              ></StarIcon>
+                            </Flex>
                           </Flex>
+                          <Text
+                            fontWeight="500"
+                            opacity={0.9}
+                            fontSize={responsive("", "sm", "md")}
+                          >
+                            {getDate(comment.createdAt)}
+                          </Text>
                         </Flex>
-                      )
-                  )}
-              </Flex>
-            </form>
+                      </Flex>
+                    )
+                  );
+                })}
+            </Flex>
           </GridItem>
         </Grid>
       </Box>
